Refactoring candidates building procedures
So that we can share some codes between candidates collecting and candidates matching
